Demo, test を examples/ に移動 `cargo --example "name"` で実行
