+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "Orinium_Browser"
+name = "orinium_browser"
 version = "0.1.0"
 edition = "2021"
