+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod engine {
+    pub mod html {
+        pub mod parser;
+        pub mod tokenizer;
+    }
+}
